Refactor DefaultHubDispatcher logging
